Combine fetch and process tabs
